move task to model
