+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Models/Task.kt
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.Tasks
+package org.firstinspires.ftc.teamcode.Models
 
 import com.qualcomm.robotcore.eventloop.opmode.OpMode
 import android.provider.SyncStateContract.Helpers.update
